requirements.txt
python-telegram-bot==20.1
selenium==4.4.3
webdriver-manager==3.8.5
